Updated current status
Affected headers 
(1)Friendly Dates in File Explorer (Build 18305)
Thank you for all of the feedback you provided on Friendly Dates in File Explorer. At this time, we’ve decided not to roll out Friendly Dates to users as part of the 19H1 release. 
(2)Re-releasing the new Japanese IME (Build 18305)
Thank you everyone who provided feedback about the new Japanese IME. We’ve decided to take some time to continue refining this experience, so that it ships with the highest possible quality, so it will not be included with 19H1, but has been reactivated for Insiders to explore in the 20H1 builds. 
(3)Updated Settings Homepage (Build 18305)
At this time the updated design is only available for Home and non-domain-joined Pro edition PCs. It’s currently in the process of rolling out to those editions – we will update this message once it’s live for everyone.